remove lambda functions
