--- netlify.toml
@@ -1,4 +1,0 @@
-[build]
-  Command = "yarn build && yarn build:lambda"
-  Functions = "lambda"
-  Publish = "build"
