+++ src/handlers/api.js
@@ -1,5 +1,7 @@
 import axios from 'axios';
-import { parseHistoricalPrices } from './parsers';
+import { parseHistoricalPrices, parseAccountAssets, parseAccountTransactions } from './parsers';
+import { infuraGetTransactionByHash, infuraGetBlockByHash } from '../handlers/infura';
+import { convertHexToString } from '../helpers/bignumber';
 import networkList from '../references/ethereum-networks.json';
 import nativeCurrencies from '../references/native-currencies.json';
 
@@ -50,6 +52,46 @@ export const apiGetMetamaskNetwork = () =>
   });
 
 /**
+ * @desc get transaction status
+ * @param  {String}   [hash = '']
+ * @param  {String}   [network = 'mainnet']
+ * @return {Promise}
+ */
+export const apiGetTransactionStatus = async (hash = '', network = 'mainnet') => {
+  try {
+    let result = await infuraGetTransactionByHash(hash, network);
+    if (!result || !result.blockNumber || !result.blockHash) return null;
+    if (result) {
+      const blockData = await infuraGetBlockByHash(result.blockHash, network);
+      result.timestamp = null;
+      if (blockData) {
+        const blockTimestamp = convertHexToString(blockData.timestamp);
+        result.timestamp = {
+          secs: blockTimestamp,
+          ms: `${blockTimestamp}000`
+        };
+      }
+    }
+    return { data: result };
+  } catch (error) {
+    throw error;
+  }
+};
+
+/**
+ * Configuration for balance api
+ * @type axios instance
+ */
+const api = axios.create({
+  baseURL: 'https://indexer.balance.io',
+  timeout: 30000, // 30 secs
+  headers: {
+    'Content-Type': 'application/json',
+    Accept: 'application/json'
+  }
+});
+
+/**
  * @desc get account balances
  * @param  {String}   [address = '']
  * @param  {String}   [network = 'mainnet']
@@ -57,14 +99,23 @@ export const apiGetMetamaskNetwork = () =>
  */
 export const apiGetAccountBalances = async (address = '', network = 'mainnet') => {
   try {
-    const { data } = await axios.get(
-      `/.netlify/functions/balance?address=${address}&network=${network}`
-    );
-    return data;
+    const { data } = await api.get(`/get_balances/${network}/${address}`);
+    const accountInfo = parseAccountAssets(data, address);
+    return accountInfo;
   } catch (error) {
     throw error;
   }
 };
+
+/**
+ * @desc get transaction data
+ * @param  {String}   [address = '']
+ * @param  {String}   [network = 'mainnet']
+ * @param  {Number}   [page = 1]
+ * @return {Promise}
+ */
+export const apiGetTransactionData = (address = '', network = 'mainnet', page = 1) =>
+  api.get(`/get_transactions/${network}/${address}/${page}`);
 
 /**
  * @desc get account transactions
@@ -78,37 +129,27 @@ export const apiGetAccountTransactions = async (
   lastTxHash = ''
 ) => {
   try {
-    let { data } = await axios.get(
-      `/.netlify/functions/transactions?address=${address}&network=${network}&lastTxHash=${lastTxHash}`
-    );
-    data = await parseHistoricalPrices(data);
-    return data;
+    let { data } = await apiGetTransactionData(address, network, 1);
+    let transactions = await parseAccountTransactions(data, address, network);
+    if (transactions.length && lastTxHash) {
+      let newTxs = true;
+      transactions = transactions.filter(tx => {
+        if (tx.hash === lastTxHash && newTxs) {
+          newTxs = false;
+          return false;
+        } else if (tx.hash !== lastTxHash && newTxs) {
+          return true;
+        } else {
+          return false;
+        }
+      });
+    }
+    transactions = await parseHistoricalPrices(transactions);
+    return transactions;
   } catch (error) {
     throw error;
   }
 };
-
-/**
- * @desc get transaction status
- * @param  {String}   [hash = '']
- * @param  {String}   [network = 'mainnet']
- * @return {Promise}
- */
-export const apiGetTransactionStatus = async (hash = '', network = 'mainnet') =>
-  axios.get(`/.netlify/functions/transaction-status?hash=${hash}&network=${network}`);
-
-/**
- * Configuration for balance proxy api
- * @type axios instance
- */
-const api = axios.create({
-  baseURL: 'https://indexer.balance.io',
-  timeout: 30000, // 30 secs
-  headers: {
-    'Content-Type': 'application/json',
-    Accept: 'application/json'
-  }
-});
 
 /**
  * @desc get ethereum gas prices
