+++ src/handlers/parsers.js
@@ -8,7 +8,9 @@ import {
   convertAmountToDisplay,
   convertAmountToDisplaySpecific,
   convertAssetAmountToNativeValue,
-  convertAssetAmountToNativeAmount
+  convertAssetAmountToNativeAmount,
+  convertStringToNumber,
+  convertAssetAmountToBigNumber
 } from '../helpers/bignumber';
 import { debounceRequest } from '../helpers/utilities';
 import { getTransactionCount } from './web3';
@@ -16,7 +18,7 @@ import { getTimeString } from '../helpers/time';
 import nativeCurrencies from '../references/native-currencies.json';
 import ethUnits from '../references/ethereum-units.json';
 import timeUnits from '../references/time-units.json';
-import { apiGetHistoricalPrices } from './api';
+import { apiGetHistoricalPrices, apiGetTransactionData } from './api';
 
 /**
  * @desc parse error code message
@@ -402,11 +404,58 @@ export const parsePricesObject = (data = null, assets = [], nativeSelected = 'US
 };
 
 /**
+ * @desc parse account assets
+ * @param  {Object} [data=null]
+ * @param  {String} [address=null]
+ * @return {Object}
+ */
+export const parseAccountAssets = (data = null, address = '') => {
+  try {
+    let assets = [...data];
+    assets = assets.map(assetData => {
+      const name =
+        assetData.contract.name !== assetData.contract.address
+          ? assetData.contract.name
+          : assetData.contract.symbol || 'Unknown Token';
+      const asset = {
+        name: name,
+        symbol: assetData.contract.symbol || '———',
+        address: assetData.contract.address || null,
+        decimals: convertStringToNumber(assetData.contract.decimals)
+      };
+      const assetBalance = convertAssetAmountToBigNumber(assetData.balance, asset.decimals);
+      return {
+        ...asset,
+        balance: {
+          amount: assetBalance,
+          display: convertAmountToDisplay(assetBalance, null, {
+            symbol: asset.symbol,
+            decimals: asset.decimals
+          })
+        },
+        native: null
+      };
+    });
+
+    assets = assets.filter(asset => !!Number(asset.balance.amount) || asset.symbol === 'ETH');
+
+    return {
+      address: address,
+      type: '',
+      assets: assets,
+      total: null
+    };
+  } catch (error) {
+    throw error;
+  }
+};
+
+/**
  * @desc parse account balances from native prices
  * @param  {Object} [account=null]
  * @param  {Object} [prices=null]
  * @param  {String} [network='']
- * @return {String}
+ * @return {Object}
  */
 export const parseAccountBalancesPrices = (account = null, nativePrices = null, network = '') => {
   let totalAmount = 0;
@@ -546,7 +595,7 @@ export const parseNewTransaction = async (
  * @param  {Object} [transactions=null]
  * @param  {String} [hash='']
  * @param  {String} [timestamp='']
- * @return {String}
+ * @return {Array}
  */
 export const parseConfirmedTransaction = (transactions = null, hash = '', timestamp = '') => {
   let _transactions = [];
@@ -561,13 +610,164 @@ export const parseConfirmedTransaction = (transactions = null, hash = '', timest
 };
 
 /**
- * @desc parse transaction historical prices
- * @param  {Object} [transactions=null]
- * @return {Object}
+ * @desc parse account transactions
+ * @param  {Object} [data=null]
+ * @param  {String} [address='']
+ * @param  {String} [networks='']
+ * @return {Array}
  */
-export const parseHistoricalPrices = async transactions => {
-  if (!transactions.length) return transactions;
+export const parseAccountTransactions = async (data = null, address = '', network = '') => {
+  if (!data || !data.docs) return [];
 
+  let transactions = await Promise.all(
+    data.docs.map(async (tx, idx) => {
+      const hash = tx._id;
+      const timestamp = {
+        secs: `${tx.timeStamp}`,
+        ms: `${tx.timeStamp}000`
+      };
+      const error = !!tx.error;
+      let interaction = false;
+      let from = tx.from;
+      let to = tx.to;
+      let asset = {
+        name: 'Ethereum',
+        symbol: 'ETH',
+        address: null,
+        decimals: 18
+      };
+      let value = {
+        amount: tx.value,
+        display: convertAmountToDisplay(tx.value, null, {
+          symbol: 'ETH',
+          decimals: 18
+        })
+      };
+      let totalGas = multiply(tx.gasUsed, tx.gasPrice);
+      let txFee = {
+        amount: totalGas,
+        display: convertAmountToDisplay(totalGas, null, {
+          symbol: 'ETH',
+          decimals: 18
+        })
+      };
+
+      const includesTokenTransfer = (() => {
+        if (tx.operations.length) {
+          const tokenTransfers = tx.operations.filter(
+            operation => operation.type === 'token_transfer'
+          );
+          if (tokenTransfers.length) {
+            return true;
+          }
+        }
+        return false;
+      })();
+
+      interaction = !includesTokenTransfer && tx.input !== '0x';
+
+      let result = {
+        hash,
+        timestamp,
+        from,
+        to,
+        error,
+        interaction,
+        value,
+        txFee,
+        native: {},
+        pending: false,
+        asset
+      };
+
+      if (includesTokenTransfer) {
+        const tokenTransfers = [];
+        if (tx.operations.length) {
+          tx.operations.forEach((transferData, idx) => {
+            const transferTx = {
+              hash: `${result.hash}-${idx + 1}`,
+              timestamp,
+              from,
+              to,
+              error,
+              interaction,
+              value,
+              txFee,
+              native: {},
+              pending: false,
+              asset
+            };
+            const name = !transferData.contract.name.startsWith('0x')
+              ? transferData.contract.name
+              : transferData.contract.symbol || 'Unknown Token';
+            transferTx.asset = {
+              name: name,
+              symbol: transferData.contract.symbol || '———',
+              address: transferData.contract.address || '',
+              decimals: transferData.contract.decimals || 18
+            };
+
+            transferTx.from = transferData.from;
+            transferTx.to = transferData.to;
+            const amount = convertAssetAmountToBigNumber(
+              transferData.value,
+              transferTx.asset.decimals
+            );
+            transferTx.value = {
+              amount,
+              display: convertAmountToDisplay(amount, null, transferTx.asset)
+            };
+            tokenTransfers.push(transferTx);
+          });
+          if (!Number(tx.value)) {
+            result = [...tokenTransfers];
+          } else {
+            result.hash = `${result.hash}-0`;
+            result = [...tokenTransfers, result];
+          }
+        }
+      }
+
+      return result;
+    })
+  );
+  let _transactions = [];
+
+  transactions.forEach(tx => {
+    if (Array.isArray(tx)) {
+      tx.forEach(subTx => {
+        _transactions.push(subTx);
+      });
+    } else {
+      _transactions.push(tx);
+    }
+  });
+
+  if (data.pages > data.page) {
+    try {
+      const newPageResponse = await apiGetTransactionData(address, network, data.page + 1);
+      const newPageTransations = await parseAccountTransactions(
+        newPageResponse.data,
+        address,
+        network
+      );
+      _transactions = [..._transactions, ...newPageTransations];
+    } catch (error) {
+      throw error;
+    }
+  }
+
+  return _transactions;
+};
+
+/**
+ * @desc parse transaction historical prices
+ * @param  {Array} [transactions=null]
+ * @return {Array}
+ */
+export const parseHistoricalPrices = async (transactions = null) => {
+  if (!transactions.length) return transactions;
+  console.log('parseHistoricalPrices transactions', transactions);
   const _transactions = await Promise.all(
     transactions.map(async (tx, idx) => {
       const timestamp = tx.timestamp ? tx.timestamp.secs : Date.now();
